+++ packages/admin/src/Resources/Pages/EditRecord.php
@@ -2,6 +2,7 @@
 
 namespace Filament\Resources\Pages;
 
+use Filament\Contracts\Page\HasFormActions;
 use Filament\Forms\ComponentContainer;
 use Filament\Notifications\Notification;
 use Filament\Pages\Actions\Action;
@@ -10,7 +11,6 @@ use Filament\Pages\Actions\ForceDeleteAction;
 use Filament\Pages\Actions\ReplicateAction;
 use Filament\Pages\Actions\RestoreAction;
 use Filament\Pages\Actions\ViewAction;
-use Filament\Pages\Contracts\HasFormActions;
 use Illuminate\Database\Eloquent\Model;
 
 /**
